requirements.txt
requests
beautifulsoup4
selenium
webdriver-manager
schedule
python-dotenv
plyer